edoc_doclet behaviour (edoc v1.5-rc0)

Standard doclet module for EDoc.

See also: edoc.

Types

command/0

-type command() :: doclet_gen() | doclet_toc().

Doclet commands.

context/0

-type context() :: #doclet_context{dir :: string(), env :: edoc:env(), opts :: [term()]}.

Context for doclets.

doclet_gen/0

-type doclet_gen() :: #doclet_gen{sources :: [string()], app :: no_app | atom(), modules :: [module()]}.

Doclet command.

doclet_toc/0

-type doclet_toc() :: #doclet_toc{paths :: [string()], indir :: string()}.

Doclet command.

Callbacks

run/2

-callback run(command(), context()) -> ok.

Doclet entrypoint.

Functions

run(Doclet_gen, Ctxt)

-spec run(edoc_doclet:command(), edoc_doclet:context()) -> ok.

Main doclet entry point. See the file edoc_doclet.hrl for the data structures used for passing parameters.

Also see edoc:layout/2 for layout-related options, and edoc:get_doc/2 for options related to reading source files.

Options:

{file_suffix, string()}

Specifies the suffix used for output files. The default value is ".html".

{hidden, boolean()}

If the value is true, documentation of hidden modules and functions will also be included. The default value is false.

{overview, edoc:filename()}

Specifies the name of the overview-file. By default, this doclet looks for a file "overview.edoc" in the target directory.

{private, boolean()}

If the value is true, documentation of private modules and functions will also be included. The default value is false.

{stylesheet, string()}

Specifies the URI used for referencing the stylesheet. The default value is "stylesheet.css". If an empty string is specified, no stylesheet reference will be generated.

{stylesheet_file, edoc:filename()}

Specifies the name of the stylesheet file. By default, this doclet uses the file "stylesheet.css" in the priv subdirectory of the EDoc installation directory. The named file will be copied to the target directory.

{title, string()}

Specifies the title of the overview-page.