edoc_doclet_markdown (edoc v1.4)

Doclet converting an edoc application to use EEP-59 and Markdown.

This doclet has to be used together with edoc_layout_chunks.

Example:

 1> edoc:application(example, [{preprocess, true}, {doclet, edoc_doclet_markdown},
       {layout, edoc_layout_chunks}]).

It will convert the overview to Markdown and any module documentation to use -doc attributes and Markdown. Any XHTML tags in the edoc documentation that are not part of the tags supported by Erlang Documentation Format will be added as HTML tags in the Markdown.

It does not delete the old edoc documentation.

See also: edoc_layout_chunks.

Functions

run(Doclet_gen, Ctxt)

-spec run(edoc_doclet:command(), edoc_doclet:context()) -> ok.

Main doclet entry point.

This doclet is tightly coupled with edoc_layout_chunks and should be used together with it.