edoc_layout behaviour (edoc v1.2)

The standard HTML layout module for EDoc. See the edoc module for details on usage.

See also: edoc.

Callbacks

module/2

-callback module(edoc:edoc_module(), _) -> binary().

Layout entrypoint.

Functions

module(Element, Options)

The layout function.

Options to the standard layout:

{index_columns, integer()}

Specifies the number of column pairs used for the function index tables. The default value is 1.

{pretty_printer, atom()}

Specifies how types and specifications are pretty printed. If the value erl_pp is specified the Erlang pretty printer (the module erl_pp) will be used. The default is to do no pretty printing which implies that lines can be very long.

{stylesheet, string()}

Specifies the URI used for referencing the stylesheet. The default value is "stylesheet.css". If an empty string is specified, no stylesheet reference will be generated.

{sort_functions, boolean()}

If true, the detailed function descriptions are listed by name, otherwise they are listed in the order of occurrence in the source file. The default value is true.

{xml_export, Module::atom()}

Specifies an xmerl callback module to be used for exporting the documentation. See xmerl:export_simple/3 for details.

See also: edoc:layout/2.

overview(E, Options)

type(E)