edoc_lib (edoc v1.3)

Utility functions for EDoc.

See also: edoc.

Types

err/0

-type err() :: eof | {missing, char()} | {line(), atom(), string()} | string().

filename/0

-type filename() :: file:filename().

line/0

-type line() :: erl_anno:line().

proplist/0

-type proplist() :: proplists:proplist().

Functions

get_doc_env(App, Modules, Options)

-spec get_doc_env(App, Modules, Options) -> edoc:env()
               when App :: atom() | no_app, Modules :: [module()], Options :: proplist().

Creates an environment data structure used by parts of EDoc for generating references, etc. See edoc:run/2 for a description of the options file_suffix, app_default and doc_path.

See also: edoc:get_doc/3, edoc_extract:source/4.

infer_module_app(Mod)

-spec infer_module_app(module()) -> no_app | {app, atom()}.

Infer application containing the given module.

It's expected that modules which are not preloaded and don't match the <app>/ebin/<mod>.beam path pattern will NOT have an app name inferred properly. no_app is returned in such cases.

write_file(Text, Dir, Name, Options)