edoc_doclet behaviour (edoc v1.0)

Standard doclet module for EDoc.

DATA TYPES

command() = doclet_gen() | doclet_toc()

Doclet commands.

context() = #doclet_context{dir=string(), env=edoc:env(), opts=[term()]}

Context for doclets.

doclet_gen() = #doclet_gen{sources=[string()], app=no_app | atom(), modules=[module()]}

Doclet command.

doclet_toc() = #doclet_toc{paths=[string()], indir=string()}

Doclet command.

See also

edoc

Functions

run(Doclet_gen:: edoc_doclet:command() , Ctxt:: edoc_doclet:context() ) -> ok

Specs

run(edoc_doclet:command(), edoc_doclet:context()) -> ok.

Main doclet entry point. See the file edoc_doclet.hrl for the data structures used for passing parameters.

Also see edoc:layout/2 for layout-related options, and edoc:get_doc/2 for options related to reading source files.

Options:

{file_suffix, string()}

Specifies the suffix used for output files. The default value is ".html".

{hidden, boolean()}

If the value is true, documentation of hidden modules and functions will also be included. The default value is false.

{overview, edoc:filename()}

Specifies the name of the overview-file. By default, this doclet looks for a file "overview.edoc" in the target directory.

{private, boolean()}

If the value is true, documentation of private modules and functions will also be included. The default value is false.

{stylesheet, string()}

Specifies the URI used for referencing the stylesheet. The default value is "stylesheet.css". If an empty string is specified, no stylesheet reference will be generated.

{stylesheet_file, edoc:filename()}

Specifies the name of the stylesheet file. By default, this doclet uses the file "stylesheet.css" in the priv subdirectory of the EDoc installation directory. The named file will be copied to the target directory.

{title, string()}

Specifies the title of the overview-page.