edoc_doclet_chunks (edoc v1.0)

Doclet generating standalone EEP-48 doc chunk files.

Section Using the EDoc API in the EDoc User's Guide shows an example of using this module.

See also

edoc_layout_chunks, shell_docs(3)

Functions

run(Doclet_gen:: edoc_doclet:command() , Ctxt:: edoc_doclet:context() ) -> ok

Specs

run(edoc_doclet:command(), edoc_doclet:context()) -> ok.

Main doclet entry point.

This doclet is tightly coupled with edoc_layout_chunks and should be used together with it.

The only option this doclet accepts is dir, which controls where the chunks subdirectory and the EEP-48 chunk files will be placed.