edoc_layout behaviour (edoc v1.4)

The standard HTML layout module for EDoc. See the edoc module for details on usage.

See also: edoc.

Types

options/0

-type options() ::
          [{index_columns, integer()} |
           {pretty_printer, atom()} |
           {stylesheet, string()} |
           {sort_functions, boolean()} |
           {xml_export, module()}].

Callbacks

module/2

-callback module(edoc:edoc_module(), _) -> binary().

Layout entrypoint.

Functions

module(Element, Options)

-spec module(Element, Options) -> term() when Element :: edoc:edoc_module(), Options :: options().

The layout function.

Options to the standard layout:

{index_columns, integer()}

Specifies the number of column pairs used for the function index tables. The default value is 1.

{pretty_printer, atom()}

Specifies how types and specifications are pretty printed. If the value erl_pp is specified the Erlang pretty printer (the module erl_pp) will be used. The default is to do no pretty printing which implies that lines can be very long.

{stylesheet, string()}

Specifies the URI used for referencing the stylesheet. The default value is "stylesheet.css". If an empty string is specified, no stylesheet reference will be generated.

{sort_functions, boolean()}

If true, the detailed function descriptions are listed by name, otherwise they are listed in the order of occurrence in the source file. The default value is true.

{xml_export, Module::atom()}

Specifies an xmerl callback module to be used for exporting the documentation. See xmerl:export_simple/3 for details.

See also: edoc:layout/2.

overview(Element, Options)

-spec overview(Element :: term(), Options :: options()) -> term().

type(Element)

-spec type(Element :: term()) -> term().