edoc_layout_chunks (edoc v1.0)

Convert EDoc module documentation to an EEP-48 docs_v1 chunk.

This layout is only expected to work with edoc_doclet_chunks. Section Using the EDoc API in the EDoc User's Guide shows an example of using this module.

This module breaks the convention stated in edoc_doclet to not rely on edoc.hrl in doclets and layouts. It uses #entry{} records directly to recover information that is not otherwise available to layouts.

DATA TYPES

beam_language() = atom()

doc() = #{doc_language() => doc_string()} | none | hidden

doc_language() = binary()

doc_string() = binary()

docs_v1() = #docs_v1{anno=erl_anno:anno(), beam_language=beam_language(), format=mime_type(), module_doc=doc(), metadata=metadata(), docs=[docs_v1_entry()]}

The Docs v1 chunk according to EEP 48.

docs_v1_entry() = {_KindNameArity::{atom(), atom(), arity()}, _Anno::erl_anno:anno(), _Signature::signature(), _Doc::doc(), _Metadata::metadata()}

A tuple equivalent to the #docs_v1_entry{} record, but with the record name field skipped.

metadata() = map()

mime_type() = binary()

signature() = [binary()]

See also

edoc_doclet_chunks, shell_docs(3)

Functions

module(Doc:: edoc:edoc_module() , Options:: proplists:proplist() ) -> binary()

Specs

module(edoc:edoc_module(), proplists:proplist()) -> binary().

Convert EDoc module documentation to an EEP-48 style doc chunk.