edoc_lib (edoc v1.4.1)

Utility functions for EDoc.

See also: edoc.

Types

proplist/0

-type proplist() :: proplists:proplist().

Functions

get_doc_env(App, Modules, Options) -> edoc:env()

Types

App = atom() | no_app

Modules = [module()]

Options = proplist()

Creates an environment data structure used by parts of EDoc for generating references, etc. See edoc:run/2 for a description of the options file_suffix, app_default and doc_path.

See also: edoc:get_doc/3, edoc_extract:source/4.

infer_module_app(Mod :: module()) -> Result

Types

Result = no_app | {app, atom()}

Infer application containing the given module.

It's expected that modules which are not preloaded and don't match the <app>/ebin/<mod>.beam path pattern will NOT have an app name inferred properly. no_app is returned in such cases.

write_file(Text, Dir, Name, Options)