edoc_lib (edoc v1.0)

Utility functions for EDoc.

DATA TYPES

proplist() = proplists:proplist()

See also

edoc

Functions

get_doc_env(App, Modules, Options) -> edoc:env()

Specs

get_doc_env(App, Modules, Options) -> edoc:env()
               when App :: atom() | no_app, Modules :: [module()], Options :: proplist().

App = atom() | no_app

Modules = [module()]

Options = proplist()

Creates an environment data structure used by parts of EDoc for generating references, etc. See edoc:run/2 for a description of the options file_suffix, app_default and doc_path.

See also: edoc:get_doc/3, edoc_extract:source/4.

infer_module_app(Mod::module()) -> no_app | {app, atom()}

Specs

infer_module_app(module()) -> no_app | {app, atom()}.

Infer application containing the given module.

It's expected that modules which are not preloaded and don't match the <app>/ebin/<mod>.beam path pattern will NOT have an app name inferred properly. no_app is returned in such cases.

write_file(Text, Dir, Name, Options) -> term()