edoc_run (edoc v1.4.1)

Interface for calling EDoc from Erlang startup options.

The following is an example of typical usage in a Makefile:

     docs:
             erl -noshell -run edoc_run application "'$(APP_NAME)'" \
               '"."' '[{def,{vsn,"$(VSN)"}}]'

(note the single-quotes to avoid shell expansion, and the double-quotes enclosing the strings).

New feature in version 0.6.9: It is no longer necessary to write -s init stop last on the command line in order to make the execution terminate. The termination (signalling success or failure to the operating system) is now built into these functions.

See also: edoc.

Types

args/0

-type args() :: [string()].

Functions

application(Args)

-spec application(args()) -> no_return().

Calls edoc:application/3 with the corresponding arguments. The strings in the list are parsed as Erlang constant terms. The list can be either [App], [App, Options] or [App, Dir, Options]. In the first case edoc:application/1 is called instead; in the second case, edoc:application/2 is called.

The function call never returns; instead, the emulator is automatically terminated when the call has completed, signalling success or failure to the operating system.

file(Args)

This function is deprecated. This is part of the old interface to EDoc and is mainly kept for backwards compatibility. The preferred way of generating documentation is through one of the functions application/1 and files/1..

-spec file(args()) -> no_return().

Calls edoc:file/2 with the corresponding arguments. The strings in the list are parsed as Erlang constant terms. The list can be either [File] or [File, Options]. In the first case, an empty list of options is passed to edoc:file/2.

The following is an example of typical usage in a Makefile:

     $(DOCDIR)/%.html:%.erl
             erl -noshell -run edoc_run file '"$<"' '[{dir,"$(DOCDIR)"}]' \
               -s init stop

The function call never returns; instead, the emulator is automatically terminated when the call has completed, signalling success or failure to the operating system.

files(Args)

-spec files(args()) -> no_return().

Calls edoc:files/2 with the corresponding arguments. The strings in the list are parsed as Erlang constant terms. The list can be either [Files] or [Files, Options]. In the first case, edoc:files/1 is called instead.

The function call never returns; instead, the emulator is automatically terminated when the call has completed, signalling success or failure to the operating system.